+++ RELEASE_NOTES.md
@@ -29,8 +29,52 @@ v9.0.0
   * Found and fixed a bug with unicode characters in the HTML parser
   * Enabled multiple set-cookie headers in the IResponse
 
-Required changes:
-1. Lorem ipsum
+Because it is no longer possible to register inner templates in the TemplateProcessor,
+   a number of methods were removed.  The removed methods are:
+   1. `renderTemplate()`
+   2. `renderTemplate(List<Map<String, String>> data, String delimiter)`
+   3. `registerInnerTemplate(String key, TemplateProcessor innerTemplate)`
+   4. `getInnerTemplate(String innerTemplateKey)`
+
+Therefore, it is now necessary to build up string output from the inside out,
+starting with the most-deeply-nested.  For example:
+
+```Java
+
+    /**
+     * Demonstrates what is necessary to build out an inner template.  In this
+     * example, we have a "ul" element representing the totality of our
+     * outer template, and then we expect to build out a list of names which is
+     * our inner template.  
+     * <br>
+     * After construction of the TemplateProcessor instances,
+     * we will first render out the internal template, then put that in the outer.
+     */
+    @Test
+    public void test_Template_SimpleInnerTemplate() {
+        // set up the templates
+        var innerTemplate = TemplateProcessor.buildProcessor("<li>{{ name }}</li>");
+        var outerTemplate = TemplateProcessor.buildProcessor("""
+                <ul>
+                    {{ inner_template_goes_here }}
+                </ul>
+                """);
+
+        // render the inner template
+        String renderedInnerTemplate = innerTemplate.renderTemplate(
+                List.of(Map.of("name", "alice"), Map.of("name", "bob")));
+
+        // merge that into the outer template, rendering the full final result
+        String finalResult = outerTemplate.renderTemplate(Map.of("inner_template_goes_here", renderedInnerTemplate));
+
+        assertEquals(finalResult, """
+                <ul>
+                    <li>alice</li>
+                    <li>bob</li>
+                </ul>
+                """);
+    }
+```
 
 v8.3.2 Dec 28, 2025
 -------------------